Remove GenerateTypeDialog_InProc type
It was pointed out that accessing automation objects can be done out-of-proc, so there isn't much need for the in-process component.
